Engine

+++ backend/src/main/java/com/tp/desafiospracticos/engine/aggregation/AggregationResult.java
@@ -2,5 +2,7 @@ package com.tp.desafiospracticos.engine.aggregation;
 
 import com.tp.desafiospracticos.engine.domain.Verdict;
 
-public record AggregationResult(int quality, Verdict suggestedVerdict) {
+import java.math.BigDecimal;
+
+public record AggregationResult(BigDecimal quality, Verdict suggestedVerdict) {
 }
